abhi/requirements.txt
streamlit>=1.28.0
openai>=1.3.0
python-dotenv>=1.0.0
langdetect>=1.0.9
ml2en>=0.1.2
gtts>=2.3.2
streamlit-mic-recorder>=0.0.7